% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{disc_api_endpoint}
\alias{disc_api_endpoint}
\title{Endpoint}
\usage{
disc_api_endpoint(endpoint_path, blanks)
}
\arguments{
\item{endpoint_path}{Discord API endpoint}

\item{blanks}{path parameters contained in a vector}
}
\description{
Endpoints contain the Discord API endpoint and the necessary path parameters.
}
\examples{
disc_api_endpoint("/channels/{}/messages", c("123456789"))
}
\author{
Dos Banaag
}
