% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{discord.client}
\alias{discord.client}
\title{Client}
\description{
This is a client that interacts with the Discord API.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-login}{\code{discord.client$login()}}
\item \href{#method-destroy}{\code{discord.client$destroy()}}
\item \href{#method-on}{\code{discord.client$on()}}
\item \href{#method-client_POST}{\code{discord.client$client_POST()}}
\item \href{#method-clone}{\code{discord.client$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-login"></a>}}
\if{latex}{\out{\hypertarget{method-login}{}}}
\subsection{Method \code{login()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.client$login(token)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-destroy"></a>}}
\if{latex}{\out{\hypertarget{method-destroy}{}}}
\subsection{Method \code{destroy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.client$destroy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-on"></a>}}
\if{latex}{\out{\hypertarget{method-on}{}}}
\subsection{Method \code{on()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.client$on(event, action)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-client_POST"></a>}}
\if{latex}{\out{\hypertarget{method-client_POST}{}}}
\subsection{Method \code{client_POST()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.client$client_POST(endpoint, body)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.client$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
