% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{discord.collection}
\alias{discord.collection}
\title{Colelction}
\description{
A collection is a vector, but with useful methods that are used throughout the library.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clear}{\code{discord.collection$clear()}}
\item \href{#method-delete}{\code{discord.collection$delete()}}
\item \href{#method-each}{\code{discord.collection$each()}}
\item \href{#method-find}{\code{discord.collection$find()}}
\item \href{#method-first}{\code{discord.collection$first()}}
\item \href{#method-push}{\code{discord.collection$push()}}
\item \href{#method-clone}{\code{discord.collection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear"></a>}}
\if{latex}{\out{\hypertarget{method-clear}{}}}
\subsection{Method \code{clear()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.collection$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\if{latex}{\out{\hypertarget{method-delete}{}}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.collection$delete(condition)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-each"></a>}}
\if{latex}{\out{\hypertarget{method-each}{}}}
\subsection{Method \code{each()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.collection$each(process)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-find"></a>}}
\if{latex}{\out{\hypertarget{method-find}{}}}
\subsection{Method \code{find()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.collection$find(condition)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-first"></a>}}
\if{latex}{\out{\hypertarget{method-first}{}}}
\subsection{Method \code{first()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.collection$first()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-push"></a>}}
\if{latex}{\out{\hypertarget{method-push}{}}}
\subsection{Method \code{push()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.collection$push(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.collection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
