% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{discord.endpoint}
\alias{discord.endpoint}
\title{Endpoint}
\description{
The endpoint class has the Discord API endpoint and the necessary path parameters.
}
\examples{
discord.endpoint$new("/channels/{}/messages", '123456789')
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{discord.endpoint$new()}}
\item \href{#method-clone}{\code{discord.endpoint$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.endpoint$new(endpoint, blanks)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint}}{Discord API endpoint}

\item{\code{blanks}}{path parameters containted in a vector}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.endpoint$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
