% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{discord.guild_manager}
\alias{discord.guild_manager}
\title{Guild Member Manager}
\description{
Guild Member Managers help store information about members in a server.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-add}{\code{discord.guild_manager$add()}}
\item \href{#method-set_client}{\code{discord.guild_manager$set_client()}}
\item \href{#method-clone}{\code{discord.guild_manager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\if{latex}{\out{\hypertarget{method-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.guild_manager$add(guild)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{guild}}{guild of the guild memer manager}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_client"></a>}}
\if{latex}{\out{\hypertarget{method-set_client}{}}}
\subsection{Method \code{set_client()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.guild_manager$set_client(input_client)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.guild_manager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
