% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{discord.message_embed}
\alias{discord.message_embed}
\title{Message Embed}
\description{
This is an embed message.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-add_field}{\code{discord.message_embed$add_field()}}
\item \href{#method-set_color}{\code{discord.message_embed$set_color()}}
\item \href{#method-set_description}{\code{discord.message_embed$set_description()}}
\item \href{#method-set_image}{\code{discord.message_embed$set_image()}}
\item \href{#method-set_thumbnail}{\code{discord.message_embed$set_thumbnail()}}
\item \href{#method-set_timestamp}{\code{discord.message_embed$set_timestamp()}}
\item \href{#method-set_title}{\code{discord.message_embed$set_title()}}
\item \href{#method-set_url}{\code{discord.message_embed$set_url()}}
\item \href{#method-get_list}{\code{discord.message_embed$get_list()}}
\item \href{#method-clone}{\code{discord.message_embed$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_field"></a>}}
\if{latex}{\out{\hypertarget{method-add_field}{}}}
\subsection{Method \code{add_field()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.message_embed$add_field(name, value, inline = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_color"></a>}}
\if{latex}{\out{\hypertarget{method-set_color}{}}}
\subsection{Method \code{set_color()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.message_embed$set_color(color)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_description"></a>}}
\if{latex}{\out{\hypertarget{method-set_description}{}}}
\subsection{Method \code{set_description()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.message_embed$set_description(new_description)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_image"></a>}}
\if{latex}{\out{\hypertarget{method-set_image}{}}}
\subsection{Method \code{set_image()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.message_embed$set_image(image)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_thumbnail"></a>}}
\if{latex}{\out{\hypertarget{method-set_thumbnail}{}}}
\subsection{Method \code{set_thumbnail()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.message_embed$set_thumbnail(thumbnail)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_timestamp"></a>}}
\if{latex}{\out{\hypertarget{method-set_timestamp}{}}}
\subsection{Method \code{set_timestamp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.message_embed$set_timestamp(isoc8601_time = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_title"></a>}}
\if{latex}{\out{\hypertarget{method-set_title}{}}}
\subsection{Method \code{set_title()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.message_embed$set_title(title)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_url"></a>}}
\if{latex}{\out{\hypertarget{method-set_url}{}}}
\subsection{Method \code{set_url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.message_embed$set_url(url)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_list"></a>}}
\if{latex}{\out{\hypertarget{method-get_list}{}}}
\subsection{Method \code{get_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.message_embed$get_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.message_embed$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
