% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{discord.text_channel}
\alias{discord.text_channel}
\title{Guild Channel}
\description{
This refers to a text channel in a server on Discord.
}
\section{Super classes}{
\code{\link[discordR:discord.channel]{discordR::discord.channel}} -> \code{\link[discordR:discord.guild_channel]{discordR::discord.guild_channel}} -> \code{discord.text_channel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{discord.text_channel$new()}}
\item \href{#method-send}{\code{discord.text_channel$send()}}
\item \href{#method-clone}{\code{discord.text_channel$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="discordR" data-topic="discord.channel" data-id="delete">}\href{../../discordR/html/discord.channel.html#method-delete}{\code{discordR::discord.channel$delete()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="discordR" data-topic="discord.channel" data-id="fetch">}\href{../../discordR/html/discord.channel.html#method-fetch}{\code{discordR::discord.channel$fetch()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.text_channel$new(guild, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{guild}}{The server of the channel}

\item{\code{data}}{Data about the channel}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-send"></a>}}
\if{latex}{\out{\hypertarget{method-send}{}}}
\subsection{Method \code{send()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.text_channel$send(content, options = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discord.text_channel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
