% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discord_colors.R
\docType{data}
\name{discord_colors}
\alias{discord_colors}
\title{Discord Color Codes}
\format{
A list with color names and their integer codes
}
\usage{
discord_colors
}
\description{
A list of common Discord embed colors
}
\examples{
discord_colors$green  # 3066993
discord_colors$red    # 15158332
}
\keyword{datasets}
