% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discordr-package.R
\docType{package}
\name{discordr-package}
\alias{discordr}
\alias{discordr-package}
\title{discordr: Send Messages and Files to Discord from R}
\description{
Provides functions to send messages, embeds, and files to Discord webhooks directly from R. Useful for monitoring long-running analyses, sending notifications, and sharing results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nbryder/discordr}
  \item Report bugs at \url{https://github.com/nbryder/discordr/issues}
}

}
\author{
\strong{Maintainer}: Nicklas Bryder \email{Nicklas.Bryder@gmail.com}

}
\keyword{internal}
