% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{discordr_setup}
\alias{discordr_setup}
\title{Discordr Setup Wizard}
\usage{
discordr_setup()
}
\value{

}
\description{
Walks the user through checking if an existing configuration exists, changing an existing configuration, or setting up a new configuration.
}
\examples{
discordr_setup()
}
