% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websocket.R
\name{DiscordrBot}
\alias{DiscordrBot}
\title{Discordr Bot}
\description{
this is the main class for the discord bot
it handles all the incoming messages and events
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DiscordrBot$new()}}
\item \href{#method-finalize}{\code{DiscordrBot$finalize()}}
\item \href{#method-start}{\code{DiscordrBot$start()}}
\item \href{#method-register_op_handler}{\code{DiscordrBot$register_op_handler()}}
\item \href{#method-register_event_handler}{\code{DiscordrBot$register_event_handler()}}
\item \href{#method-send_payload}{\code{DiscordrBot$send_payload()}}
\item \href{#method-get_ws_endpoint}{\code{DiscordrBot$get_ws_endpoint()}}
\item \href{#method-discord_get_api}{\code{DiscordrBot$discord_get_api()}}
\item \href{#method-discord_post_api}{\code{DiscordrBot$discord_post_api()}}
\item \href{#method-clone}{\code{DiscordrBot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscordrBot$new(token, intent = GUILD_MESSAGES)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{token}}{The discord bot token}

\item{\code{intent}}{the intent(s) that the bot should access.
Discordr defines constant for the possible intents
Closes websocket connection
Start the bot}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscordrBot$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start"></a>}}
\if{latex}{\out{\hypertarget{method-start}{}}}
\subsection{Method \code{start()}}{
starts the bot and enters in an infinite later loop
if in a interactive session returns as the later event loop will be run while R is idle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscordrBot$start()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-register_op_handler"></a>}}
\if{latex}{\out{\hypertarget{method-register_op_handler}{}}}
\subsection{Method \code{register_op_handler()}}{
register op handler
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscordrBot$register_op_handler(op, callback)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{op}}{one of the opcodes from the discord websocket
https://discord.com/developers/docs/topics/opcodes-and-status-codes}

\item{\code{callback}}{a function that would be called with the event data
register event handlers}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-register_event_handler"></a>}}
\if{latex}{\out{\hypertarget{method-register_event_handler}{}}}
\subsection{Method \code{register_event_handler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscordrBot$register_event_handler(event_type, callback)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event_type}}{one of the events from the discord websocket
https://discord.com/developers/docs/topics/gateway#commands-and-events}

\item{\code{callback}}{a function that would be called with the event data
send payload directly to the websocket}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-send_payload"></a>}}
\if{latex}{\out{\hypertarget{method-send_payload}{}}}
\subsection{Method \code{send_payload()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscordrBot$send_payload(opcode, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{opcode}}{websocket OP CODE}

\item{\code{data}}{list that represents data to be sent in payload
return the endpoint of Websocket connections
general GET request to discord}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ws_endpoint"></a>}}
\if{latex}{\out{\hypertarget{method-get_ws_endpoint}{}}}
\subsection{Method \code{get_ws_endpoint()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscordrBot$get_ws_endpoint()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-discord_get_api"></a>}}
\if{latex}{\out{\hypertarget{method-discord_get_api}{}}}
\subsection{Method \code{discord_get_api()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscordrBot$discord_get_api(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{the path from the discord base api.
general POST request to discord}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-discord_post_api"></a>}}
\if{latex}{\out{\hypertarget{method-discord_post_api}{}}}
\subsection{Method \code{discord_post_api()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscordrBot$discord_post_api(path, body)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{the path from the discord base api.}

\item{\code{body}}{a list that will the body of the http request.
It will automatically encoded in json
create a new websocket and connect to it
switcher for op code
Switch according to received events
hello is the first message sent by discord on the websocket
need to setup heartbeat and intent
connect to discord sending intents and
The session is invalid for discord
need to reconnect again
resumes previous sessions after they are interrupted
reconnect socket if asked by discord
save the ready event}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscordrBot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
