% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{enable_console_logging}
\alias{enable_console_logging}
\title{Enable logging for the bot in console}
\usage{
enable_console_logging(level = 20)
}
\arguments{
\item{level}{logging level}
}
\description{
Enable logging for the bot in console
}
