% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{enable_file_logging}
\alias{enable_file_logging}
\title{Enable logging for the bot in a file}
\usage{
enable_file_logging(level = 20, file = "discordr-logs.log")
}
\arguments{
\item{level}{logging level}

\item{file}{file where logs should be saved}
}
\description{
Enable logging for the bot in a file
}
