% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_setup.R
\name{export_discord_connection}
\alias{export_discord_connection}
\title{Export Discord Connections}
\usage{
export_discord_connection(conn, filepath, append = TRUE)
}
\arguments{
\item{conn}{Connection Object to Export}

\item{filepath}{Filepath to new or existing connections file}

\item{append}{Optional; Assumed to be TRUE to add connection objects, otherwise will overwrite existing file}
}
\description{
Export Discord Connections
}
\examples{
\dontrun{
conn_obj <- create_discord_connection(webhook = 'https://google.com', username = 'test')
export_discord_connection(conn_obj, 'discord_connections.csv')
}
}
