% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_setup.R
\name{get_default_discord_connection}
\alias{get_default_discord_connection}
\title{Get Default Discord Connnection}
\usage{
get_default_discord_connection()
}
\value{
DiscordR Connection Object
}
\description{
Retreives Default Connection Object from R Options if set
}
\examples{
\dontrun{
get_default_discord_connection()
}
}
