% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_default_discord_username}
\alias{get_default_discord_username}
\title{Get Default Discordr Username}
\usage{
get_default_discord_username(verbose = TRUE)
}
\arguments{
\item{verbose}{Return detailed messages on if username is currently set.}
}
\value{
Currently set discordr username environment variable; Will return an empty string and message if no username has been set.
}
\description{
Obtains the currently set default username or returns an error if it not set within the current environment. If username is not set, use \code{\link{set_default_discord_username}} to set default environment username.
}
\examples{
\dontrun{
get_default_discord_username()
}
}
