% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_discordr_username}
\alias{get_discordr_username}
\title{Get Default Discordr Username}
\usage{
get_discordr_username()
}
\value{

}
\description{
Obtains the currently set default username or returns an error if it not set within the current environment. If username is not set, use \code{\link{set_discordr_username}} to set default environment username.
}
\examples{
get_discordr_username()
}
