% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_discordr_webhook}
\alias{get_discordr_webhook}
\title{Get Default Discordr Webhook}
\usage{
get_discordr_webhook(verbose = TRUE)
}
\arguments{
\item{verbose}{Return detailed messages on if webhook is currently set.}
}
\value{
Currently set webhook address environment variable; Will return an empty string and message if no webhook address is set
}
\description{
Obtains the currently set default username or returns an error if it not set within the current environment. If a default webhook is not set, Use \code{\link{set_discordr_webhook}} to set default environment webhook.
}
\examples{
\dontrun{
get_discordr_webhook()
}
}
