% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{get_token}
\alias{get_token}
\title{Get a bot token}
\usage{
get_token()
}
\value{
Character vector. The bot token that has been set by \code{\link[=set_token]{set_token()}}
}
\description{
Gets the bot token to be used in API calls that require authorization.
}
\author{
Erez Shomron
}
