% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_setup.R
\name{import_discord_connections}
\alias{import_discord_connections}
\title{Import Discord Connections From File}
\usage{
import_discord_connections(filepath)
}
\arguments{
\item{filepath}{Path to Discord Connections File}
}
\value{
List of Connection Objects
}
\description{
Import Discord Connections From File
}
\examples{
\dontrun{
import_discord_connections('test.csv')
}
}
