% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{message}
\alias{message}
\title{Message}
\usage{
message(client, data, channel)
}
\arguments{
\item{client}{The client that initiated the guild}

\item{data}{Data about the guild}

\item{channel}{The channel of the message}
}
\description{
This is a message.
}
\author{
Dos Banaag
}
