% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest-api.R
\name{paste_url}
\alias{paste_url}
\title{Paste two url components}
\usage{
paste_url(...)
}
\arguments{
\item{...}{elements to paste}
}
\description{
shorthand to call paste with sep="/", useful when creating url paths
}
\examples{
paste_url("rest", "api", "path")

}
