% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_functions.R
\name{send_console}
\alias{send_console}
\title{Send Console Output}
\usage{
send_console(
  ...,
  conn = get_default_discord_connection(),
  filename = tempfile(pattern = "discordr")
)
}
\arguments{
\item{...}{A single or set of expressions to be evaluated for console output}

\item{conn}{Discord Connection Object containing Webhook and Username}

\item{filename}{Alternative path to route console output; defaults to tempfile}
}
\value{
None
}
\description{
This functions accepts an expression whose console output you would like to send to the specified discord channel
}
\examples{
\dontrun{
send_console(2 + 2)
}
}
\seealso{
\code{\link{send_message}}, \code{\link{send_file}}, \code{\link{send_plot_code}}, \code{\link{send_current_ggplot}}
}
