% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_functions.R
\name{send_current_ggplot}
\alias{send_current_ggplot}
\title{Send Current GGPlot}
\usage{
send_current_ggplot(
  conn = get_default_discord_connection(),
  filename = tempfile(pattern = "discordr", fileext = ".png")
)
}
\arguments{
\item{conn}{Discord Connection Object containing Webhook and Username}

\item{filename}{Optional - Filepath indicating where to save image; Provide to manually override the temporary directory and filename}
}
\value{
None
}
\description{
Send images of GGplot which are currently shown in the plots pane of RStudio or elsewhere.
}
\examples{
\dontrun{
send_current_ggplot()
}
}
\seealso{
\code{\link{send_plot_code}}, \code{\link{send_file}}, \code{\link{send_message}}, \code{\link{send_console}}
}
