% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{send_current_plot}
\alias{send_current_plot}
\title{Send Current Plot}
\usage{
send_current_plot(conn = get_default_discord_connection(),
  filename = tempfile(pattern = "discordr", fileext = ".png"))
}
\arguments{
\item{filename}{Optional - Filepath indicating where to save image; Provide to manually override the temporary directory and filename}

\item{username}{Username to use for sender of message, defaults to environment set username}

\item{webhook}{Webhook to which the message should be sent, defaults to environment set webhook}
}
\value{
None
}
\description{
Sends the last plot displayed within the plots tab in RStudio. This function will display both ggplot and base graphics
with a lower resolution than only ggplot graphics. In order to save and send the file, a random name for the image will
be generated and saved temporarily.
}
\examples{
\dontrun{
send_current_plot()
}
}
\seealso{
\code{\link{send_current_ggplot}}, \code{\link{send_file}}, \code{\link{send_message}}, \code{\link{send_console}}
}
