% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest-api-tools.R
\name{send_embed}
\alias{send_embed}
\title{send a discord embed}
\usage{
send_embed(embed, channel_id, bot)
}
\arguments{
\item{embed}{a discordr_embed object}

\item{channel_id}{id of the channel where the embed should be sent}

\item{bot}{an instance of DiscordrBot}
}
\description{
send a discord embed
}
\seealso{
\link{new_embed}
}
