% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_functions.R
\name{send_file}
\alias{send_file}
\title{Send File}
\usage{
send_file(filename, conn = get_default_discord_connection())
}
\arguments{
\item{filename}{filepath and filename of the file to be sent}

\item{conn}{Discord Connection Object containing Webhook and Username}
}
\value{
None
}
\description{
Send file from the given filename to the username and webhook provided
}
\examples{
\dontrun{
send_file('image.jpg')
}
}
\seealso{
\code{\link{send_file}}, \code{\link{send_plot_code}}, \code{\link{send_current_ggplot}}, \code{\link{send_console}}
}
