% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{send_message}
\alias{send_message}
\title{Send Message}
\usage{
send_message(message, conn = get_default_discord_connection())
}
\arguments{
\item{message}{Character string of message to send}

\item{username}{Username to use for sender of message, defaults to environment set username}

\item{webhook}{Webhook to which the message should be sent, defaults to environment set webhook}
}
\value{
None
}
\description{
Sends a message using the username provided to the channel of the webhook provided
}
\examples{
\dontrun{
send_message("Hello World!")
}

}
\seealso{
\code{\link{send_file}}, \code{\link{send_current_plot}}, \code{\link{send_current_ggplot}}, \code{\link{send_console}}
}
