% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{send_robject}
\alias{send_robject}
\title{Send R Objects}
\usage{
send_robject(
  ...,
  filename = tempfile(pattern = "discordr", fileext = ".RData"),
  username = get_discordr_username(),
  webhook = get_discordr_webhook()
)
}
\arguments{
\item{...}{Single or Multiple R Objects to be contained within a single RData file}

\item{filename}{Default is a random string saved in the temporary directory; change this if you would like the RData file to be human-readable and in a different location.}

\item{username}{Username to use for sender of message, defaults to environment set username}

\item{webhook}{Webhook to which the message should be sent, defaults to environment set webhook}
}
\value{
None
}
\description{
Send R Objects
}
\examples{
\dontrun{
x <- c(1,2,3,4,5)
y <- matrix(rep(0, 4), rows = 2, cols = 2)
send_robject(x, y, filename = 'test_data.RData')
}
}
