% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{send_tex}
\alias{send_tex}
\title{Send Rendered Latex Images}
\usage{
send_tex(tex_string, filename = tempfile(pattern = "discordr"),
  density = 250, username = get_discordr_username(),
  webhook = get_discordr_webhook())
}
\arguments{
\item{tex_string}{Character string of compileable latex code. Ensure you are using double slashes ('\\') for commands.}

\item{filename}{Default is a random string saved in the temporary directory; change this if you would like the RData file to be human-readable and in a different location.}

\item{density}{Density of latex image to be saved. Default is 250.}

\item{username}{Username to use for sender of message, defaults to environment set username}

\item{webhook}{Webhook to which the message should be sent, defaults to environment set webhook}
}
\value{
None
}
\description{
In order to use this function, you will need to install the texPreview package manually.
}
