% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_setup.R
\name{set_default_discord_connection}
\alias{set_default_discord_connection}
\title{Set Default Discord Connection}
\usage{
set_default_discord_connection(conn)
}
\arguments{
\item{conn}{Discord Connection Object}
}
\description{
Set Default Discord Connection
}
\examples{
\dontrun{
conn_obj <- create_discord_connection(webhook = 'https://google.com', username = 'test')
set_default_discord_connection(conn_obj)
}
}
