% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_default_discord_username}
\alias{set_default_discord_username}
\title{Set Default Discordr Username}
\usage{
set_default_discord_username(username)
}
\arguments{
\item{username}{Username to be used as default discord name}
}
\value{
None
}
\description{
Sets a default username to be used for discord communication. Use \code{\link{get_default_discord_username}} to check currently set default username.
}
\examples{
\dontrun{
set_default_discord_username("dataman")
}
}
