% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_discordr_username}
\alias{set_discordr_username}
\title{Set Default Discordr Username}
\usage{
set_discordr_username(username)
}
\arguments{
\item{username}{Username to be used as default discord name}
}
\value{

}
\description{
Sets a default username to be used for discord communication. Use \code{\link{get_discordr_username}} to check currently set default username.
}
\examples{
set_discordr_username("dataman")
}
