% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{token_get}
\alias{token_get}
\title{Get a bot token}
\usage{
token_get()
}
\value{
Character vector. The bot token that has been set by \code{\link[=token_set]{token_set()}}
}
\description{
Gets the bot token to be used in API calls that require authorization.
}
\author{
Erez Shomron
}
