% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user}
\alias{user}
\title{User}
\usage{
user(client, data)
}
\arguments{
\item{client}{Discord client}

\item{data}{initial data collected about the user}
}
\description{
Creates a user
}
\author{
Dos Banaag
}
