% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_setup.R
\name{create_discord_connection}
\alias{create_discord_connection}
\title{Create Discord Connection Object}
\usage{
create_discord_connection(
  webhook_string,
  username = get_default_discord_username(verbose = FALSE),
  server_name = "",
  channel_name = "",
  set_default = FALSE
)
}
\arguments{
\item{webhook_string}{Webhook URL for sending discord comments}

\item{username}{Defaults to \code{get_default_discord_username()}}

\item{server_name}{Optional - Used for discriminating webhooks in a human readable format}

\item{channel_name}{Optional - Used for discriminating webhooks in a human readable format}

\item{set_default}{Optional - Set created connection as the default connection}
}
\value{
DiscordR connection object containing provided information
}
\description{
Create Discord Connection Object
}
\examples{
\dontrun{
conn_obj <- create_discord_connection(webhook = 'https://google.com', username = 'test')
}
}
