% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{send_console}
\alias{send_console}
\title{Send Console Output}
\usage{
send_console(..., conn = get_default_discord_connection(),
  filename = tempfile(pattern = "discordr"))
}
\arguments{
\item{...}{A single or set of expressions to be evaluated for console output}

\item{filename}{Optional - Save console output to a file, defaults to a temporary file}

\item{username}{Username to use for sender of message, defaults to environment set username}

\item{webhook}{Webhook to which the message should be sent, defaults to environment set webhook}
}
\value{
None
}
\description{
This functions accepts an expression whose console output you would like to send to the specified discord channel
}
\examples{
\dontrun{
send_console(2 + 2)
}
}
\seealso{
\code{\link{send_message}}, \code{\link{send_file}}, \code{\link{send_current_plot}}, \code{\link{send_current_ggplot}}
}
