% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{send_current_ggplot}
\alias{send_current_ggplot}
\title{Send Current Plot (ggplot version)}
\usage{
send_current_ggplot(
  username = get_discordr_username(),
  webhook = get_discordr_webhook(),
  filename = tempfile(pattern = "discordr", fileext = ".png")
)
}
\arguments{
\item{username}{Username to use for sender of message, defaults to environment set username}

\item{webhook}{Webhook to which the message should be sent, defaults to environment set webhook}

\item{filename}{Optional - Filepath indicating where to save image; Provide to manually override the temporary directory and filename}
}
\value{

}
\description{
This function works identical to \code{\link{send_current_plot}} except that it appears to provide a higher resolution and only work for plots
constructed using a ggplot workflow.
}
\examples{
send_current_ggplot()

}
\seealso{
\code{\link{send_current_plot}}, \code{\link{send_file}}, \code{\link{send_message}}, \code{\link{send_console}}
}
