% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{send_file}
\alias{send_file}
\title{Send File}
\usage{
send_file(
  filename,
  username = get_discordr_username(),
  webhook = get_discordr_webhook()
)
}
\arguments{
\item{filename}{filepath and filename of the file to be sent}

\item{username}{Username to use for sender of message, defaults to environment set username}

\item{webhook}{Webhook to which the message should be sent, defaults to environment set webhook}
}
\value{

}
\description{
Send file from the given filename to the username and webhook provided
}
\examples{
send_file('image.jpg')

}
\seealso{
\code{\link{send_file}}, \code{\link{send_current_plot}}, \code{\link{send_current_ggplot}}
}
