% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest-api.R
\name{send_message}
\alias{send_message}
\title{Send discord message}
\usage{
send_message(message, channel_id)
}
\arguments{
\item{message}{message content}

\item{channel_id}{id of the channel}
}
\description{
sends a discord message on a given channel id
}
