% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest-api-tools.R
\name{send_message}
\alias{send_message}
\title{send discord message}
\usage{
send_message(message, channel_id, bot)
}
\arguments{
\item{message}{message content}

\item{channel_id}{id of the channel}

\item{bot}{an instance of DiscordrBot}
}
\description{
a discord message on a given channel id
}
