% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_functions.R
\name{send_message}
\alias{send_message}
\title{Send Message}
\usage{
send_message(message, conn = get_default_discord_connection())
}
\arguments{
\item{message}{Character string of message to send}

\item{conn}{Discord Connection Object containing Webhook and Username}
}
\value{
None
}
\description{
Sends a message using the username provided to the channel of the webhook provided
}
\examples{
\dontrun{
send_message("Hello World!")
}

}
\seealso{
\code{\link{send_file}}, \code{\link{send_plot_code}}, \code{\link{send_current_ggplot}}, \code{\link{send_console}}
}
