% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_functions.R
\name{send_tex}
\alias{send_tex}
\title{Send Rendered Latex Images}
\usage{
send_tex(
  tex_string,
  filename = tempfile(pattern = "discordr"),
  density = 250,
  conn = get_default_discord_connection()
)
}
\arguments{
\item{tex_string}{Character string of compileable latex code. Ensure you are using double slashes ('\\') for commands.}

\item{filename}{Default is a random string saved in the temporary directory; change this if you would like the RData file to be human-readable and in a different location.}

\item{density}{Density of latex image to be saved. Default is 250.}

\item{conn}{Discord Connection Object containing Webhook and Username}
}
\value{
None
}
\description{
In order to use this function, you will need to install the texPreview package manually.
}
\examples{
\dontrun{
tex_string <- "$\\\\int^a_b \\\\frac{1}{3}x^3 dx$"
send_tex(tex_string, conn = conn_obj)
}
}
